Implemented propagated opacity and added CC_PROPAGATE_OPACITY option in ccConfig.h.
If enabled, cocos2d uses transitive opacity. All classes calculate its cumulative opacity from self opacity and parent's opacity. For example, if you setup a CCLayerRGB with an opacity of 127 (50%) and add a children CCSprite with an opacity of 127,
it draws like having an opacity of 63 (25%).

For performance reasons, it's recommended to disable it if you don't need the feature.
To enable set it to 1. Use 0 to disable it. Enabled by default