Add fragment shader for alpha testing.
OpenGLES 2.0 does not support glAlphaTest or glAlphaFunc.
This change set adds a simple fragment shader to the default set of cocos2d shaders
that performs the equivalent of a glAlphaTest with by comparing the incoming alpha value
with a specified uniform float, using the equivalent of a (greater-than-or-equal-to)
'>=' comparison.  This produces the same results as:

glAlphaFunc(GL_GREATER, <value>)
glEnable(GL_ALPHA_TEST)

However, due to the use of shaders, the configuration is slightly different, since
for cocos2d, use of the alpha test shader must specified on a per-drawing-node basis.


Former-commit-id: f91ac14a5134ac9023498fd74def6d1c9582889a