Add support for easily caching texture instances from any file.
In some situations, texture files are stored on disk with names that do not match
the content.  For example, a generic game 'resource' system that supports network
loadable assets might require the use of hashes for filenames.  In this situation,
it is necessary to provide params to override the defaults that are normally extracted
from a filename.

This change set provides support for caching instances of custom subclasses of CCTexture2D
with params specifying the cache key, and the "file format" string. The key can be any string,
and is used to cache the texture under a known key. The fileFormat string is currently used to
override the file extension extracted from the filename. Currently, these additional params are
only implemented for methods used to load & cache PVR textures.