Mac project: compiler AMC as part of cocos2d lib only. Closes #3
In iOS project AMC failed to compile as a part of static lib.
After fixing compiler errors it compiled ok as part
only of cocos2d lib, without adding it to test targets.

Also it made it possible to use it the same way on
mac.

Probably there was some optimization on Mac, that
implicitly removed AMC sources from cocos2d lib target,
cause it was impossible to build it as a part of lib.