Performance dispatcher test
These files are for  performance dispatcher test. If needed performance
dispatcher test can be added to the existing 6 performance test targets.
The results of test are very good. There is a substantial improvement
in the speed for all of the operations from 30% to 1500% depending on
the API operation.
The biggest gain of speed occurs for reordering priorities of the
delegates and for removal of the delegates.

TEST RESULTS FOR iPod 1gen iOS 3.1.3 (in debug mode)
Test setup: 200 delegates (100 Standard delegates and 100 targeted
delegates)
// Advantages of the new API:
1 a - FAST ADD (0x0023cd00) : avg time, 33.023032ms
1 a - FAST ADD (0x0023cd00) : avg time, 32.198968ms
1 a - FAST ADD (0x0023cd00) : avg time, 31.350551ms
// SPEED GAIN > 33.0%
2a FAST REMOVAL (0x002564c0) : avg time, 1.504204ms
2a FAST REMOVAL (0x002564c0) : avg time, 1.578540ms
2a FAST REMOVAL (0x002564c0) : avg time, 1.489650ms
// SPEED GAIN > 1500%
2b Fast Compound removal (0x0024c260) : avg time, 1.958303ms
2b Fast Compound removal (0x0024c260) : avg time, 2.066753ms
2b Fast Compound removal (0x0024c260) : avg time, 2.045931ms
// SPEED GAIN > 1500%
5a QSort (0x002564c0) : avg time, 134.083539ms
5a QSort (0x002564c0) : avg time, 134.018829ms
5a QSort (0x002564c0) : avg time, 134.094870ms
// SPEED GAIN > 65%
5b MSort (0x0023cd00) : avg time, 149.510898ms
5b MSort (0x0023cd00) : avg time, 149.723841ms
5b MSort (0x0023cd00) : avg time, 150.293651ms
// SPEED GAIN > 45%
5c UltraFast(delayed)Reordering (0x002222c0) : avg time, 24.026498ms
5c UltraFast(delayed)Reordering (0x002222c0) : avg time, 23.920655ms
5c UltraFast(delayed)Reordering (0x002222c0) : avg time, 24.230325ms
// SPEED GAIN > 820%

Of course even if one does not need speed gains the main advantage of
new dispatcher code is that it works correctly in the touch callbacks.