More permissive texture clamping check and assert that a shader program is set during CCNode's draw setup.
-[CCTexture2D setTexParameters:] currently asserts that textures are
either POT or use clamping.  This is too restrictive and doesn't permit
things like REPEAT on POT dimensions while the other dimension isn't
POT.

CC_NODE_DRAW_SETUP() enables the node's shader program.  Usage of this
macro indicates that the node is about to draw something.  This draw
operation requires that a shader program is set on the node, otherwise
the resulting draw operation is undefined (it's indetermined which
shader program will be used for it).  Therefore, it's safe to assume
that no shader program being set on the node is a bug and we can assert
during the function that a shader program is actually set to catch these
bugs fast.


Former-commit-id: 557b5603e5c35cbb2809ae66596efc8135be7b04