Added useTrimmedFrameForAnchorPoint property to CCSprite.
When set to YES this property makes the trimmed frame the frame of reference for the contentSize and the anchorPoint. When using trimmed frames with different offsets and same original sizes, the anchor point will be off. Or when using different trimmed frames with different offsets. 

This is useful for animations that use the trimmed frame as frame of reference for transformations or expect the anchor point to remain the same between frames with different offsets. 

To test the new behavior, I've included two new sprite tests.