Proper disabling of events during transitions
Touch events were disabled when transition objects were initialized, and reenabled when they successfully finished. Initialization may be done in advance, so it's too early to disable events here. Successful finishing is not guaranteed, as transition scenes may be popped from the director.

The correct place for these disable/enable events would be the onEnter and onExit methods of the CCTransition class.