Fixes build on Xcode 4.2
Cocos2d 2.1 introduces several build errors with Xcode 4.2 because NSLineBreakBy______ and NSTextAlignment______ are not defined. This uses the preprocessor to define them if the Xcode version is less than 4.4.

This is primarily for developers who are stuck on Mac OS X 10.6 because their hardware is too old (e.g Macbook Pro 1st gen). For those, the newest Xcode version available is 4.2.