Fix memory leak in CCRenderTexture
Since we were only keeping track of the most recently allocated GL FBO and depth buffer,
we were leaking resources when rendering multiple passes. The fix is to keep track of all
allocated FBOs and depth buffers and delete these in addition to the CCTexture objects in
the destroy method.