Continue CCEffect stacking
The bulk of this change is the movement certain responsibilities from CCEffect subclasses and into
CCEffectNode. Before this, multi-pass effects were responsible for rendering the children of CCEffectNode
into an FBO as their first pass and then rendering the results of the last pass into the displayed frame buffer.
Now that we're connecting effects together, this responsibility can't fall on each effect since an one effect
might be taking its input from (or writing its output to) another.

CCEffectNode now renders its children into an FBO prior to the first effect's first pass and then draws the
results of the last effect's last pass into the displayed frame buffer. All CCEffects now have a sampler uniform
called cc_PreviousPassTexture that can be used for reading the results of the previous pass.

It is possible that we will want to give CCEffects the ability to override the pre-render and post-render steps
that are now owned by CCEffectNode but we don't have a use case for that yet.