Set the cc_MainTexture uniform equal to the texture at the beginning of the current effect
It had been the texture at the beginning of all effects but that's less useful. For the most part
an effect should really only depend on its input texture, the textures corresponding to intermediate
render passes, and any uniforms it declares. Reaching all the way back to the texture at the beginning
of an effect chain seems less meaningful to me though we may end up finding a use for this in the
future.