Set the position+color+texture shader in CCEffectNode's post-rendering composite phase
The original implementation of post-render composite phase didn't set the shader which meant
it was left as whatever the last effect's last render pass had set it to. This hadn't caused any visible
problems yet but it would have eventually and it also meant we were running a more expensive
shader than we needed to be.