Change CCEffectPixellate to use a pixel size parameter instead of pixel scale
This addresses Viktor's feedback in issue #670 that pixel scale was unclear (Which it was
especially since it wasn't even correctly named. It was really the uv step value not a scale
factor at all). Now the user can specify the size of the pixel blocks and the effect does the
math to convert this into uv step values.