Fix CCEffectNode anchor point issues when no effect is present
Oleg found a bug in which the CCEffectNode's anchor point yielded different results when a CCEffect was present
versus when it wasn't. The problem was that effects change the node's sprite's position and this was not being reset.
The fix is to set the effect node's sprite's anchor point and position to (0,0) during the compositing pass. This way the
sprite is at the same position as the containing effect node.