CCEffectRenderer refactor - Remove multiple render pass support from CCRenderTexture.
As we move to CCEffectRenderer which is shared by CCSprite and CCEffectNode, we need to
move render pass management in there. This mean the intermediate texture / FBO array needs
to come out of CCRenderTexture since it needs to go where the passes go.

This change temporarily make CCEffectNode function just like CCRenderTexture by disabling
the execution of the render passes (because they have no textures to store intermediate results).