CCEffectRenderer refactor - Initial implementation of CCEffectRenderer
This moves the main effect render loop from CCEffectNode into CCEffectRenderer. Management
of the FBOs and textures for each render pass is also moved into the renderer. We're currently
leaking FBOs and textures with each draw but that will be fixed next (or at least very soon if not
next).