CCEffectRenderer refactor - Fix initialization of cc_MainTexture
I would like for this fix to be temporary but for now it gets the glow effect working again. Because
of the way we manipulate CCEffectNode's sprite (the internal one, not the one that the effect is
being applied to) I was tripping over some interactions between dirtiness testing, sprite.texture,
and the cc_MainTexture uniform. An issue like this has come up at least once before and this
fix also just feels like a band-aid so I'd like to address it differently (i.e. better) in the future.