First pass at CCEffects applied directly to CCSprites
Sprite contains an effect renderer that renders to a temporary texture which is then applied
to the sprite during the final compositing render pass. There is a big hack here in which a
temporary effect stack is allocated during each draw to be passed to the renderer. This will
go away once CCEffectStack is itself a subclass of CCEffect (kind of like the relationship
between CCAction and CCActionSequence).