Subclass CCEffectStack from CCEffect - Start moving render pass ownership into CCEffect
Until now CCEffectRenderPass was a temporary structure that was created and filled out
by the renderer and handed to an effect's begin, update, and end methods at draw time.
Once CCEffectStack is subclassed from CCEffect, ownership of the render passes needs
to move into CCEffect to avoid a bunch of special case logic in the renderer (i.e. we don't
want the renderer to care about the specific subclass of CCEffect, we just want it to be able
to iterate over the current effect's passes).

This is the first part of that ownership change and it adds render pass creation code to each
CCEffect subclass. These passes are not used yet though. That change is next.