Subclass CCEffectStack from CCEffect - Renderer conversion
The renderer now takes a CCEffect as its input instead of a CCEffectStack and internally
it no longer differentiates between CCEffect and CCEffectStack. It just iterates over all the
passes of the supplied effect executing them accordingly.

This change does temporarily break effects that refer to pass outputs other than the previous
one (CCEffectGlow for example) but a fix for that is coming soon.