CCEffect stitching - WIP - Start experimenting with effect input declarations
Instead of putting texture loads from cc_PreviousPassTexture directly into a an effect function's
source, provide a mechanism for function's to declare the inputs they expect. With the declaration
of each input, include a snippet of code that can be used to load the input value from the previous
pass's texture if the function is the first one in a chain of functions that are stitched together. If a
function is first in a chain, it's loader snippet is inserted into the stitched shader, if function is later
in the chain, its input is passed from the output of the previous function.