CCEffect stitching - Supply the render pass as an argument to the pass blocks
Originally the pass was captured by each block when the block was defined but this doesn't
work with effect stitching. When effects are stitched together, multiple passes are merged
into one pass but the blocks for the original passes are maintained. Since a block now needs
to reference a different pass than the one that created it, the pass must be passed in at call
time.