CCEffect stitching - WIP - Start working on name mangling
To prevent identifier name collisions when stitching shader source together we need to prefix
identifiers with unique strings. For now this string is <effect_debug_string>_<effect_index_in_stack>.
The stitcher is generating valid GLSL but a few items still need to be addressed:

- Varying values
- Uniform name strings in effect update blocks. The name strings are hardcoded but the stitcher
is changing them in the GLSL source. We need a way to update the names that the blocks use
when updating the uniform dictionary.
- Handling default CCEffect uniforms like cc_PreviousPassTexture.