CCEffect stitching - Finish name mangling
Effects now have a translation dictionary which can be used to map original uniform names
to their mangled variants. When a set of effects are stitched together, the stitcher creates a new
translation table and updates the stitched effects with it. The current implementation has the side
effect that one CCEffect object cannot be used in multiple effect stacks but I think that's acceptable.
At least until we see that there is a need for the sharing of effects between stacks.