CCEffect stitching - Simplify stitch list creation
Instead of doing stitch list creation and stitching in one pass, create the stitch lists
first and then do the stitching in two separate passes. Since all of this stuff is prone
to becoming complicated and ugly as it develops, it's best to keep it simpler where
possible. Creating the stitch lists first and stitching them later makes it easier to debug
these two different phases of the process.