CCEffect stitching - Remove recursion in CCEffectStack prepareForRendering
Stitching stacks of stacks through recursion was not only slow but broken in one case. Rework
the code so that it first builds a flat list of effects within the stack (and its substacks). From there
it builds the stitch lists which is not only simpler and less bug prone but faster since it no longer
has to stitch together the results of other intermediate stitch operations.