CCEffectRefraction - Remove some magic numbers
Instead of the environment being supplied as a CCTexture it is now a CCSprite. This allows the
effect to account for the position of the environment within the scene when computing texture
coordinates. From the environment sprite, compute the offset and scale to map from screen space
fragment positions to environment texture space. This accounts for environment sprite position and
scale but I've not tested rotation or non-default texture rects. That means they're broken.