CCEffectRefraction - Improve the mapping from screen space coordinates to environment coordinates
The previous code would not work with all transformations that could be applied to the environment sprite
(changes to the texture rect, rotation, etc). This code constructs the affine transform from screen space
coordinates to the environment sprite's texture coordinates so it handles all possible transforms of the
environment.

There is still one issue remaining with my handling of the tangent and binormal vectors. The fix for that
is coming up.