CCEffectStack - Remove addEffect and removeEffect
Viktor's feedback was to follow the model of CCActionSequence and make CCEffectStack immutable.
Even though it means it will not be used for now, I'm keeping the passesDidChange notification method.
It's possible that changes to the parameters of stacked effects will dirty the flattened pass list and the
ability to bubble the dirty state up to the root of a stack tree will still be necessary.