CCEffectsTest - Update refraction test to use a CCRenderTexture for the environment
Using a render texture instead of just a sprite allows for dynamic environments which
makes the effect that much more interesting to look at. Note that getting the texture
coordinate math to work requires setting the anchor point of the render texture's internal
sprite to (0, 0) from its default value of (0.5, 0.5).