CCSprite - Refactor setTextureCoords
Change it from the instance method setTextoreCoords to the class method textureCoordsForTexture... All of the
instance variables that it used to read are now passed in as parameters and the result is returned to the caller.
It is the caller's responsibility to copy the results to the appropriate texture coordinate set in the sprite's vertices.
With these changes the method can now be used for computing both the base texture coordinates and the normal
map texture coordinates.