CCEffect - Rework shader uniform management
The node (sprite or effect node) to which an effect is attached is now the owner of the uniform dictionary
that is used during effect rendering. When an effect is attached to a node, the effect's uniforms are copied
into the uniform dictionary of the sprite / node to which it is being attached. Later during rendering, the
effect renderer retrieves the uniform dictionary from the node and hands it to the render passes as they
are executed.

Because effect stitching can modify uniform names, the uniform dictionary needs to be updated if
prepareForRendering is called and does any work on an effect stack. This means that the renderer no
longer calls prepareForRendering and it's instead the caller's responsibility to call it when it is necessary.
The renderer will assert that an effect is ready for rendering before it uses it.

The effect (hah) of all of this is that the node effectively (hahaha) becomes the local scope for uniform
variables and the uniform dictionary maintained by each effect is a template for the uniforms that need
to be created in that local scope. We can't reverse this relationship (copying a node's uniforms into the
effect's uniform dictionary) because that would prevent us from attaching an effect to more than one sprite
which is a desired goal of the system.