CCEffects - Fix compiler warning and add an assert for an unsupported situation
We don't currently support effects with multiple inputs* so assert that we're not getting
them. Because of this limitation we only need one temporary variable when stitching effects
together. Fix a compiler warning where we had a loop over all inputs that we don't need.

*Note: In this context, the word "input" means some color value that is potentially the output
of a preceding effect in an effect stack.