CCEffectRefraction - Tweak the refraction math a bit
While experimenting with reflection I came up with some modifications to the refraction effect that
yield better results (IMO). I've also come up with a mapping of min and max refraction values that
yield a nice range of visual results (again IMO). 1.0 results in maximum magnification of the environment
image, 0.0 is no change, and -1.0 results in maximum shrinking of the environment image. Please
note that this is based on experimentation and what looks good and I'm not trying to make this an
accurate optical simulation. For instance you might ask:

Q: Why does the refraction value scale the output of the refract() function instead of being used as the
IOR ratio input to refract()?
A: I tried both and I found that the former yields better looking results than the latter.

Q: Where do the min and max refraction values come from?
A: I experimented with a wide range of values and found that these yield pleasing results while
minimizing unexpected behavior.