CCEffectReflection / Glass - Fake reflection math tweaks.
Increase the power that (1 - nDotV) is raised to and then scale the result
back into the range [0..1]. Since nDotV is proportional to the cosine of the
angle between N and V, raising it to a power > 1 has the effect of flattening
out the center of the curve which looks better to me when applied to a
spherical normal map.

I should have been scaling (1-nDotV) all along since it was ending up in the
range [0..2] which resulted in excessive compression of the environment map
image at the silhouette edges of the normal map. This only gets worse when
raising to a power so the scale is even more important here.