CCEffects - Add debug label to CCEffectRenderPass
While profiling the effects code I noticed that construction of each render pass's debug
label via [NSString stringWithFormat] was pretty high on the list of expensive code paths.
Instead of constructing this dynamically during rendering it's now created when the render
passes are created.

This results in a 14% performance boost in the effect performance test with the saturation
effect.