CCEffect - Fix vertex shaders when stacking / stitching
The outputs of stacked vertex shaders were still adding together instead of feeding one into
the other as pipeline stages. This meant that odd things were happening with vertex values
(xy position would stay the same but the w coord would change) which lead to unexpected
results when using gl_FragCoord in reflection and refraction math.

The fix is to treat vertex shader inputs and outputs the same as we do with fragment shader
inputs and outputs. To this end I've refactored buildEffectShader so vertex and fragment
shaders are handled by a common code path.