CCEffects - Fix assertion failures when texture + normal map are missing
Under certain circumstances, the effect renderer could allocate a render target
with 0 x 0 dimensions which resulted in a GL error and assertion failure. Those
circumstances were:

- Sprite has no texture
- Sprite has no normal map
- An unstitchable effect stack is attached to the sprite

This is now fixed in the following ways:

- The effect renderer's render target dimensions now come from sprite.contentSize
not just sprite.texture.contentSize. This means that if the sprite has a normal map but
not a texture, the resulting size will come from the normal map instead of being 0 x 0 as
it was before.

- If the computed render target size is still 0 x 0, it is bumped to 1 x 1.

- An assertion on 0 x 0 dimensions is added to the render target allocation code
so this sort of error is caught higher up in the stack in the future.