CCEffectBloom - Fix threshold math
When applying the luminance threshold, don't check the luminance of the blurred result pixel and instead
check the luminance of the inputs to the blur operation. If the luminance of an input falls below the
threshold just use black as the input instead. Because the threshold happens before the blur, the
boundaries of the thresholded areas are still nicely blurred. Doing the threshold after the blur results
in hard boundaries that don't achieve the effect we want.

Note that the addition of the luminance calculation and the step() function does have a negative
performance impact. We're going with this implementation for now but this is an argument for
supporting multiple shaders / implementations per-effect. With multiple implementations, we could
select one that does not pay the performance penalty for the threshold check if it's not needed.