CCSprite - Fix texture coordinate setup from normal map sprite frame
If a sprite has a normal map but not a texture, we want to set up the texture rect from the
normal map. This was mostly implemented but because [setTextureRect: ...] referenced self.texture
it was broken. Instead of referencing self.texture in this case you need to reference the normal
map texture.

Fix this by refactoring the guts of [setTextureRect: ...] into [setTextureRect: forTexture: ...] which
takes the texture to use as a parameter instead of always using self.texture. Calls to the original
[setTextureRect: ...] just call through to [setTextureRect: forTexture: ...] supplying self.texture as
the texture parameter.