CCEffects - Fix mixing of sprite color / opacity with effects
Some effects handled colors and opacity properly, some did not. And none of them did the
right thing if multiple effects were stacked together without stitching. Color would either be
multiplied in multiple times or not at all.

This change introduces a new code snippet for effect function inputs that is used if the effect
function is the first one in a stack. Otherwise, the alternative snippet is used. The initial snippet
for all effects multiplies the fragment color by the texture and the alternate snippet just fetches
the texture.

In plain english this means that the input to an effect stack is the sprite's texture times its color
and we insure that the texture times color multiply does not happen again in the stack.