CCEffectNode - More parent relative sizing fixes
Keep track of the last allocated content size and compare this to the current content size.
If the two are different, destroy and reallocate the associated texture and FBO. This fixes
cases where the effect node's size is parent relative (and does not change) but the parent's
size does change.