Added reschedule:interval method to CCNode.
Added reschedule:interval method to CCNode.m. When you call the reschedule:interval: method using a selector that is already scheduled this method will update the interval.  It may be useful to use this method at certain times rather than using the  schedule:interval: method (which will unschedule and then reschedule an already scheduled method).

If you call reschedule:interval method and pass in a selector that was never scheduled, this method will schedule the selector for the first time.