CCEffect - Add support for effect padding
Fix clipping of effects that expand the visual footprint of the node they're applied to (blur for example).
The user can now specify horizontal and vertical padding that increases the size of the rendered
geometry but adjusts the texture coordinates such that the sprite image stays the same size. By
adding this padding, effects like blur no longer get clipped off by the edge of the sprite geometry.

Unfortunately, for sprites from sprite sheets, individual effects must also be modified to avoid
fetching outside the sprite's footprint within the sheet. Because GLES does not support the
clamp-to-border wrap mode, this must be handled within the shaders.