CCEffectRenderer - Fix texture coordinates for intermediate render passes from sprite sheets
We had been using the sprite's original texture coordinates for all render passes which will not
work with sprite sheets. When reading from an intermediate render target we need the texture
coordinates to be lower left (0,0) and upper right (1,1) and this is not going to be the case when
a sprite draws its image from a sprite sheet.

Fix this problem by overwriting the sprite texture coordinates with the required values when reading
from a sprite sheet.