CCEffects - Fixes for environment mapping effects and CCRenderTexture
The previous method of finding the transform from an effect to its environment via the
root of the scene did not work when the node and environment were children of a
CCRenderTexture or when they were being manually rendered to a CCRendertTexture
(with direct calls to begin / visit / end). Instead find the transform from the node to its environment
via their closest common ancestor.

This approach is compatible with CCRenderTexture so long as that common ancestor
is a descendant of the render texture (either explicitly via the node graph or implicitly via the
rendering of a node subtree via manual begin / visit / end). If instead the render texture
is a descendant of the common ancestor then there is no way to always find a valid
transform from the node to its environment.