CCEffectLighting - Directional light support
For point lights we compute the light vector from each vertex position to the light position
and interpolate this across the sprite. For directional lights we set the light vector to be the
inverse of the light's direction (because light direction comes "from" the light but we want
a vector "to" the light). This is the same for all vertices so the direction is constant across
the sprite (and therefore independent of light position).