CCEffectLighting + CCLightNode - Light depth
Even though cocos2d scenes, nodes, and sprites are two dimensional, it is unavoidable
that some lighting math involves a third dimension. Because normal maps define normal
vectors in three dimensions, we need a way to specify a virtual Z value for our lights or
otherwise all lights will appear to be coplanar with the sprites they are lighting. In this case
all lighting would appear to be from the side (or above or below) and would be unpleasantly
harsh looking.

This change adds support for a light Z coordinate by adding a depth property (I've avoided
using the name z or Z since this name is already used for render order sorting).