CCEffects - [CCEffect prepareForRendering] -> [CCEffect prepareForRenderingWithSprite]
Since prepareForRendering is responsible for dynamic shader generation in response to property changes,
the lighting shader is dependent on the presence or absence of a normal map, and the normal map lives
on the sprite, the affected sprite needs to be supplied to prepareForRendering.  That was not being done
in the past. It is now.