CCEffectLighting + CCLightNode - Non-linear falloff support
In addition to the cutoff radius, add a half radius value to CCLightNode. This is the distance at which
a light's intensity has fallen off to 50% of its maximum value. The half radius plus the cutoff radius is
used to construct a piecewise linear falloff function from (0,1) to (half,0.5) to (cutoff,0).