CCEffectLighting - First pass at automatic light collection and light groups
Instead of requiring the user to manually add light nodes to their light effects, automatically
collect all lights in a scene and let the effects query for the closest ones. Which lights
contribute to which effects is controlled similarly to collisions. Lights and effects indicate
which groups they are in and if there is an intersection between a light's groups and an
effect's groups then the light contributes to the effect.

Note: Right now the query is only returning the first N lights not the closest N lights.