CCEffects - Refactor CCEffectRenderPass
CCEffectRenderPass was serving two purposes: 1) To pass static information to the effect renderer and 2) To receive information
from the renderer and convey it to the begin, update, and end blocks.  This dual purpose "I'm an input and an output" business was
starting to get a little hard to keep track of so I'm busting it up into two classes: CCEffectRenderPass is the static stuff that the effect
is passing to the renderer. CCEffectRenderPassInputs is the stuff that the renderer is passing to the effect's blocks.