CCEffects - Refactor effects into CCEffect and CCEffectImpl
This is something we've thought about for a while as a way to enable different
implementations as optimizations but I'm doing it now in preparation for adding
Metal support.

With this change, the functions, uniforms, varyings, and render passes all move
into CCEffectImpl subclass (or really an effect specific subclass of CCEffectImpl)
while the public initializers, convenience constructors, and properties all stay in
the CCEffect subclass.

For now, initializing an effect instance immediately creates the corresponding
effect implementation instance and the implementation does not change for the
lifetime of the effect.  This need not always be the case though and it is possible
to switch the implementation dynamically based on effect property values.

For example, the bloom effect currently has a conditional in its shader based on the
luminance threshold value. If the threshold is 0.0 or 1.0 though, we could select
a different implementation where the condition is replaced with the hardcoded
pass or fail path as an optimization.