CCEffectLighting - Lighting collection fixes
- Don't filter the light list based on distance. I tried this optimization but its causing some
problems so I'm removing it for now.
- Provide an interface for clearing the light collection's group name list. This is mostly just
used for the tests since I want to be able to reset the group names / masks with each test.
- Add a method to CCLightCollection that returns the ambient color sum of all lights matching
a supplied group mask. This is necessary since we want lights to contribute their ambient
color even if they are not in the set of closest K lights that is used by the shader.