Fix CCLabelBMFonts.
This class had accumulated a lot of technical debt.  The removal the tag
method of addressing nodes had been dealt with in an idiosyncratic way
that resulted in skip, justSkipped and a lot of strange addressing logic
that made font aligning and line wrapping impossible to debug.

The logic to recycle the CCSprite which were used for the individual
font characters was done by traversing the child nodes using a
replacement getChildByTag method that really did not handle what was
being attempted with the alignment and line-wrapping.

This patch:

 - subclasses CCSprite to create a CCBMFontCharacter
    - knows what character it represents
    - has a pointer to the subsequent character

 - iterating through all active characters no longer uses the children
   array, and avoids the issue of nodes that may currently be
   non-visible, waiting to be re-used.

 - The line breaking on words and in the middle of very long words
   now works as expected.

 - Alignment for left, center and right works now

 - fixes the CCLabelBMFont debug draw which was not working

 - Adds tests to the Cocos2D tests target to allow testing
   wrapping and aligned labels which were not previously tested.
*

Conflicts:
	cocos2d-tests.xcodeproj/project.pbxproj
	cocos2d/CCLabelBMFont.m