Edge case fixes, get characters test.
There were a few broken combinations of fonts and alignment - edge case
bugs, now fixed.

Also: added tests for the characterSpritesForRange: method, which
fetches a list of the sprites corresponding to a specified range in the
string.  This can be used for example to color text or to apply effects.