Fixes for bmfont scaled fonts.
Support was added for "FooBar.bmfont" type fonts, where the heierarchy
contains a FooBar.fnt and FooBar.png texture for each of the device
resolutions.  The .fnt and .png textures are typically scaled versions
of the bit-map font, for example exported at 96, 48, 48 and 24 point
from the tool (eg GlyphDesigner).

These fonts have a content scale value in the .fnt file that exposed a
bug in the last fix-cclabelbmfont fixes, where the commonHeight value
needed to be scaled by this content scale amount when being used to
advance the line Y position for a line-feed.

Also there were issues where the soft-line feeds inserted needed to be
removed again (by regenerating the font chars) when the width of the
label is dynamically changed.

Added a set of tests and an example font to capture these issues.  The
font is from an Open Source font project and its license is in the file
COPYING - it has explicit exclusions to state that the GPL here does not
cause containing projects to be virally affected by using the font.