Removing the setting of _userObject to nil
In the cleanup method the _userObject property was being set to nil - appears to be a leftover from some earlier AnimationManager code from the commit history. _userObject should be left alone by the core engine, as by definition it is something for user code to do whatever it wants with.