Fixing the color space lookup for initWithUIColor
CCColor initWithUIColor was using self.GCColor to work out the color space to use rather than the passed in color. Self at this point is un-initialised so wouldn't be giving good information and since change addb32507676c47cfc1624d66ec44893c9d16cbd it breaks the further use of CGColor on the object as that will have been lazy allocated, probably to pure transparent black - 0/0/0/0.

Previous to that commit this method would have caused a leak of a CGColor object.